requirements.txt
crewai==0.5.0
google-generativeai>=0.3.0
tiktoken>=0.5.1
python-dotenv>=1.0.0
rich>=13.7.0
langchain>=0.0.350
langchain-google-genai>=0.0.6
